+++ clients/admin-ui/src/features/data-discovery-and-detection/action-center/fields/page.tsx
@@ -17,7 +17,7 @@ import { Key, useEffect, useState } from "react";
 import { DebouncedSearchInput } from "~/features/common/DebouncedSearchInput";
 import FixedLayout from "~/features/common/FixedLayout";
 import { getErrorMessage } from "~/features/common/helpers";
-import { useSearch } from "~/features/common/hooks";
+import { useAlert, useSearch } from "~/features/common/hooks";
 import { ACTION_CENTER_ROUTE } from "~/features/common/nav/routes";
 import PageHeader from "~/features/common/PageHeader";
 import { useAntPagination } from "~/features/common/pagination/useAntPagination";
@@ -26,12 +26,17 @@ import {
   useClassifyStagedResourcesMutation,
   useGetMonitorConfigQuery,
   useGetMonitorFieldsQuery,
+  useIgnoreMonitorResultAssetsMutation,
 } from "~/features/data-discovery-and-detection/action-center/action-center.slice";
+import {
+  usePromoteResourcesMutation,
+  useUpdateResourceCategoryMutation,
+} from "~/features/data-discovery-and-detection/discovery-detection.slice";
 import { DiffStatus } from "~/types/api";
 import { isErrorResult } from "~/types/errors";
 
 import { MonitorFieldFilters } from "./MonitorFieldFilters";
-import MonitorFieldListItem from "./MonitorFieldListItem";
+import renderMonitorFieldListItem from "./MonitorFieldListItem";
 import MonitorTree from "./MonitorTree";
 import { RESOURCE_STATUS, useMonitorFieldsFilters } from "./useFilters";
 
@@ -78,6 +83,7 @@ const ActionCenterFields: NextPage = () => {
     monitor_config_id: monitorId,
   });
   const [selectedNodeKeys, setSelectedNodeKeys] = useState<Key[]>([]);
+  const [selectedFields, setSelectedFields] = useState<string[]>([]);
   const { data: fieldsDataResponse } = useGetMonitorFieldsQuery({
     monitor_config_id: monitorId,
     size: pageSize,
@@ -92,6 +98,43 @@ const ActionCenterFields: NextPage = () => {
   const toast = useToast();
   const [classifyStagedResourcesMutation] =
     useClassifyStagedResourcesMutation();
+  const [updateResourceCategoryMutation] = useUpdateResourceCategoryMutation();
+  const [ignoreMonitorResultAssetsMutation] =
+    useIgnoreMonitorResultAssetsMutation();
+  const { errorAlert } = useAlert();
+  const [promoteResources] = usePromoteResourcesMutation();
+
+  const handleSetDataUses = async (dataUses: string[], urn: string) => {
+    const mutationResult = await updateResourceCategoryMutation({
+      monitor_config_id: monitorId,
+      staged_resource_urn: urn,
+      user_assigned_data_categories: dataUses,
+    });
+
+    if (isErrorResult(mutationResult)) {
+      errorAlert(getErrorMessage(mutationResult.error));
+    }
+  };
+
+  const handleIgnore = async (urn: string) => {
+    const mutationResult = await ignoreMonitorResultAssetsMutation({
+      urnList: [urn],
+    });
+
+    if (isErrorResult(mutationResult)) {
+      errorAlert(getErrorMessage(mutationResult.error));
+    }
+  };
+
+  const handlePromote = async (urns: string[]) => {
+    const mutationResult = await promoteResources({
+      staged_resource_urns: urns,
+    });
+
+    if (isErrorResult(mutationResult)) {
+      errorAlert(getErrorMessage(mutationResult.error));
+    }
+  };
 
   const handleClassifyStagedResources = async () => {
     const result = await classifyStagedResourcesMutation({
@@ -182,7 +225,18 @@ const ActionCenterFields: NextPage = () => {
                     Filter
                   </Button>
                 </Dropdown>
-                <Dropdown menu={{ items: [] }}>
+                <Dropdown
+                  menu={{
+                    items: [
+                      {
+                        key: "promote",
+                        onClick: () => handlePromote(selectedFields),
+                        label: "Promote",
+                      },
+                    ],
+                  }}
+                  disabled={selectedFields.length <= 0}
+                >
                   <Button
                     type="primary"
                     icon={<Icons.ChevronDown />}
@@ -196,7 +250,20 @@ const ActionCenterFields: NextPage = () => {
             <List
               dataSource={fieldsDataResponse?.items}
               className="overflow-scroll"
-              renderItem={MonitorFieldListItem}
+              renderItem={(props) =>
+                renderMonitorFieldListItem({
+                  ...props,
+                  selected: selectedFields.includes(props.urn),
+                  onSelect: (key, selected) =>
+                    selected
+                      ? setSelectedFields([...selectedFields, key])
+                      : setSelectedFields(
+                          selectedFields.filter((val) => val !== key),
+                        ),
+                  onSetDataUses: handleSetDataUses,
+                  onIgnore: handleIgnore,
+                })
+              }
             />
             <Pagination
               {...paginationProps}