+++ clients/admin-ui/src/features/data-discovery-and-detection/action-center/tables/cells/DiscoveredAssetActionsCell.tsx
@@ -8,10 +8,11 @@ import {
 import { truncate } from "lodash";
 import { useRouter } from "next/router";
 
+import { useFeatures } from "~/features/common/features/features.slice";
 import { getErrorMessage, isErrorResult } from "~/features/common/helpers";
 import { SYSTEM_ROUTE } from "~/features/common/nav/routes";
 import { errorToastParams, successToastParams } from "~/features/common/toast";
-import { ConsentStatus, DiffStatus } from "~/types/api";
+import { DiffStatus } from "~/types/api";
 import { StagedResourceAPIResponse } from "~/types/api/models/StagedResourceAPIResponse";
 
 import {
@@ -19,26 +20,25 @@ import {
   useIgnoreMonitorResultAssetsMutation,
   useRestoreMonitorResultAssetsMutation,
 } from "../../action-center.slice";
-import { DiscoveryErrorStatuses } from "../../constants";
 import { ActionCenterTabHash } from "../../hooks/useActionCenterTabs";
 import { SuccessToastContent } from "../../SuccessToastContent";
+import hasConsentComplianceIssue from "../../utils/hasConsentComplianceIssue";
 
 interface DiscoveredAssetActionsCellProps {
   asset: StagedResourceAPIResponse;
   onTabChange: (tab: ActionCenterTabHash) => Promise<void>;
   showComplianceIssueDetails?: (
     stagedResource: StagedResourceAPIResponse,
-    status: ConsentStatus,
   ) => void;
-  showWarningForConsentIssues?: boolean;
 }
 
 export const DiscoveredAssetActionsCell = ({
   asset,
   onTabChange,
   showComplianceIssueDetails,
-  showWarningForConsentIssues,
 }: DiscoveredAssetActionsCellProps) => {
+  const { flags } = useFeatures();
+  const { assetConsentStatusLabels: isConsentStatusFlagEnabled } = flags;
   const [addMonitorResultAssetsMutation, { isLoading: isAddingResults }] =
     useAddMonitorResultAssetsMutation();
   const [ignoreMonitorResultAssetsMutation, { isLoading: isIgnoringResults }] =
@@ -68,9 +68,8 @@ export const DiscoveredAssetActionsCell = ({
   const truncatedAssetName = truncate(name || "", { length: 50 });
 
   // Check if the consent status is an error type
-  const isErrorStatus = consentAggregated
-    ? DiscoveryErrorStatuses.includes(consentAggregated)
-    : false;
+  const showConsentComplianceWarning =
+    hasConsentComplianceIssue(consentAggregated);
 
   const handleAdd = async () => {
     const result = await addMonitorResultAssetsMutation({
@@ -128,10 +127,7 @@ export const DiscoveredAssetActionsCell = ({
   };
 
   const handleViewComplianceDetails = () => {
-    showComplianceIssueDetails?.(
-      asset,
-      consentAggregated ?? ConsentStatus.UNKNOWN,
-    );
+    showComplianceIssueDetails?.(asset);
   };
 
   // TODO [HJ-369] update disabled and tooltip logic once the categories of consent feature is implemented
@@ -165,7 +161,7 @@ export const DiscoveredAssetActionsCell = ({
           >
             Ignore
           </Button>
-          {isErrorStatus && showWarningForConsentIssues && (
+          {showConsentComplianceWarning && isConsentStatusFlagEnabled && (
             <Button
               data-testid="view-compliance-details-btn"
               size="small"