+++ clients/admin-ui/src/features/data-discovery-and-detection/discovery-detection.slice.ts
@@ -186,7 +186,7 @@ const discoveryDetectionApi = baseApi.injectEndpoints({
           },
         )}`,
       }),
-      invalidatesTags: ["Discovery Monitor Results"],
+      invalidatesTags: ["Discovery Monitor Results", "Monitor Field Results"],
     }),
     muteResources: build.mutation<any, BulkResourceActionQueryParams>({
       query: ({ staged_resource_urns }) => ({
@@ -206,10 +206,13 @@ const discoveryDetectionApi = baseApi.injectEndpoints({
           { arrayFormat: "none" },
         )}`,
       }),
-      invalidatesTags: ["Discovery Monitor Results"],
+      invalidatesTags: ["Discovery Monitor Results", "Monitor Field Results"],
     }),
     updateResourceCategory: build.mutation<
-      any,
+      Array<{
+        urn: string;
+        data_uses?: string[] | null;
+      }>,
       ChangeResourceCategoryQueryParam
     >({
       query: (params) => ({
@@ -224,7 +227,7 @@ const discoveryDetectionApi = baseApi.injectEndpoints({
         ],
       }),
 
-      invalidatesTags: ["Discovery Monitor Results"],
+      invalidatesTags: ["Discovery Monitor Results", "Monitor Field Results"],
     }),
   }),
 });